onebusawaykt-kotlin-core/src/main/kotlin/com/onebusawaykt/api/models/StopReportProblemRetrieveResponse.kt
// File generated from our OpenAPI spec by Stainless.

package com.onebusawaykt.api.models

import com.fasterxml.jackson.annotation.JsonAnyGetter
import com.fasterxml.jackson.annotation.JsonAnySetter
import com.fasterxml.jackson.annotation.JsonProperty
import com.fasterxml.jackson.databind.annotation.JsonDeserialize
import com.onebusawaykt.api.core.ExcludeMissing
import com.onebusawaykt.api.core.JsonField
import com.onebusawaykt.api.core.JsonMissing
import com.onebusawaykt.api.core.JsonValue
import com.onebusawaykt.api.core.NoAutoDetect
import com.onebusawaykt.api.core.toUnmodifiable
import java.util.Objects

@JsonDeserialize(builder = StopReportProblemRetrieveResponse.Builder::class)
@NoAutoDetect
class StopReportProblemRetrieveResponse
private constructor(
    private val code: JsonField<Long>,
    private val currentTime: JsonField<Long>,
    private val text: JsonField<String>,
    private val version: JsonField<Long>,
    private val additionalProperties: Map<String, JsonValue>,
) {

    private var validated: Boolean = false

    private var hashCode: Int = 0

    fun code(): Long = code.getRequired("code")

    fun currentTime(): Long = currentTime.getRequired("currentTime")

    fun text(): String = text.getRequired("text")

    fun version(): Long = version.getRequired("version")

    @JsonProperty("code") @ExcludeMissing fun _code() = code

    @JsonProperty("currentTime") @ExcludeMissing fun _currentTime() = currentTime

    @JsonProperty("text") @ExcludeMissing fun _text() = text

    @JsonProperty("version") @ExcludeMissing fun _version() = version

    @JsonAnyGetter
    @ExcludeMissing
    fun _additionalProperties(): Map<String, JsonValue> = additionalProperties

    fun validate(): StopReportProblemRetrieveResponse = apply {
        if (!validated) {
            code()
            currentTime()
            text()
            version()
            validated = true
        }
    }

    fun toBuilder() = Builder().from(this)

    override fun equals(other: Any?): Boolean {
        if (this === other) {
            return true
        }

        return other is StopReportProblemRetrieveResponse &&
            this.code == other.code &&
            this.currentTime == other.currentTime &&
            this.text == other.text &&
            this.version == other.version &&
            this.additionalProperties == other.additionalProperties
    }

    override fun hashCode(): Int {
        if (hashCode == 0) {
            hashCode =
                Objects.hash(
                    code,
                    currentTime,
                    text,
                    version,
                    additionalProperties,
                )
        }
        return hashCode
    }

    override fun toString() =
        "StopReportProblemRetrieveResponse{code=$code, currentTime=$currentTime, text=$text, version=$version, additionalProperties=$additionalProperties}"

    companion object {

        fun builder() = Builder()
    }

    class Builder {

        private var code: JsonField<Long> = JsonMissing.of()
        private var currentTime: JsonField<Long> = JsonMissing.of()
        private var text: JsonField<String> = JsonMissing.of()
        private var version: JsonField<Long> = JsonMissing.of()
        private var additionalProperties: MutableMap<String, JsonValue> = mutableMapOf()

        internal fun from(stopReportProblemRetrieveResponse: StopReportProblemRetrieveResponse) =
            apply {
                this.code = stopReportProblemRetrieveResponse.code
                this.currentTime = stopReportProblemRetrieveResponse.currentTime
                this.text = stopReportProblemRetrieveResponse.text
                this.version = stopReportProblemRetrieveResponse.version
                additionalProperties(stopReportProblemRetrieveResponse.additionalProperties)
            }

        fun code(code: Long) = code(JsonField.of(code))

        @JsonProperty("code")
        @ExcludeMissing
        fun code(code: JsonField<Long>) = apply { this.code = code }

        fun currentTime(currentTime: Long) = currentTime(JsonField.of(currentTime))

        @JsonProperty("currentTime")
        @ExcludeMissing
        fun currentTime(currentTime: JsonField<Long>) = apply { this.currentTime = currentTime }

        fun text(text: String) = text(JsonField.of(text))

        @JsonProperty("text")
        @ExcludeMissing
        fun text(text: JsonField<String>) = apply { this.text = text }

        fun version(version: Long) = version(JsonField.of(version))

        @JsonProperty("version")
        @ExcludeMissing
        fun version(version: JsonField<Long>) = apply { this.version = version }

        fun additionalProperties(additionalProperties: Map<String, JsonValue>) = apply {
            this.additionalProperties.clear()
            this.additionalProperties.putAll(additionalProperties)
        }

        @JsonAnySetter
        fun putAdditionalProperty(key: String, value: JsonValue) = apply {
            this.additionalProperties.put(key, value)
        }

        fun putAllAdditionalProperties(additionalProperties: Map<String, JsonValue>) = apply {
            this.additionalProperties.putAll(additionalProperties)
        }

        fun build(): StopReportProblemRetrieveResponse =
            StopReportProblemRetrieveResponse(
                code,
                currentTime,
                text,
                version,
                additionalProperties.toUnmodifiable(),
            )
    }
}
